Moves 'extensions' to 'modules'
