+++ src/Dialog/modules/useButtonRole.test.js
@@ -1,6 +1,9 @@
 /* eslint-disable max-len */
 import Dialog from '..';
-import { ManageTabIndex } from '.';
+import { UseButtonRole } from '.';
+import { events } from '../../../.jest/events';
+
+const { keydownEnter, keydownSpace } = events;
 
 // Set up our document body
 document.body.innerHTML = `
@@ -32,27 +35,26 @@ document.body.innerHTML = `
 
 const controller = document.querySelector('[aria-controls="dialog"]');
 
-const dialog = new Dialog(controller, { extensions: [ManageTabIndex] });
+const dialog = new Dialog(controller, { modules: [UseButtonRole] });
 
-test('The Dialog target\'s interactive children are out of tab index when hidden', () => {
-  dialog.interactiveChildElements.forEach((link) => {
-    expect(link.getAttribute('tabindex')).toEqual('-1');
-  });
-
-  dialog.expanded = true;
-  expect(dialog.expanded).toBe(true);
-  dialog.interactiveChildElements.forEach((link) => {
-    expect(link.getAttribute('tabindex')).toBeNull();
-  });
-
+beforeEach(() => {
   dialog.expanded = false;
-  expect(dialog.expanded).toBe(false);
-  dialog.interactiveChildElements.forEach((link) => {
-    expect(link.getAttribute('tabindex')).toEqual('-1');
-  });
+});
 
-  dialog.destroy();
-  dialog.interactiveChildElements.forEach((link) => {
-    expect(link.getAttribute('tabindex')).toBeNull();
-  });
+test('Should use the button role on the controller', () => {
+  expect(controller.getAttribute('role')).toBe('button');
+  expect(controller.getAttribute('tabindex')).not.toBe('0');
+});
+
+test('The Return key and Spacebar activate the Dialog target', () => {
+  // Ensure the Dialog is closed.
+  expect(dialog.expanded).toBe(false);
+
+  // Enter.
+  controller.dispatchEvent(keydownEnter);
+  expect(dialog.expanded).toBe(true);
+
+  // Spacebar.
+  controller.dispatchEvent(keydownSpace);
+  expect(dialog.expanded).toBe(true);
 });
