+++ src/Disclosure/README.md
@@ -66,12 +66,12 @@ _**`target`**_ `HTMLElement`
 
 > **Note** Full event details documented at [`src/README`](../).
 
-## Extensions
+## Modules
 
 ```jsx
 import Disclosure, { ManageTabIndex } from 'aria-components/disclosure';
 
-const disclosure = new Disclosure(element, { extensions: [ManageTabIndex] });
+const disclosure = new Disclosure(element, { modules: [ManageTabIndex] });
 ```
 
 ## References
