+++ src/Tablist/modules/automaticActivation.js
